Clean up UCTNode, split off UCTNodePointer.
Move UCTNodePointer code to separate files, clean up coding style.
